Title: 트리에서 공 떨어뜨리기, Time: 3035ms, Memory: 23MB, Status: Time Limit Exceed - CodeTree

+++ 240218/트리에서 공 떨어뜨리기/drop-a-ball-in-a-tree.java
@@ -19,7 +19,7 @@ public class Main {
             right[i] = r;
         }
 
-        int K = Integer.parseInt(br.readLine());
+        long K = Long.parseLong(br.readLine());
         while(K-- > 0){
             int curIdx = 1;
             while(true){
